requirements.txt
fastapi>=0.115.0
uvicorn>=0.25.0
jinja2>=3.0
python-multipart>=0.0.7
pdfminer.six>=20221105
PyPDF2>=3.0.0